airflow_end/requirements.txt
pandas
pandasql
sqlalchemy
pyspark==3.1.2
apache-airflow==2.2.3
apache-airflow-providers-apache-cassandra==2.0.0
apache-airflow-providers-apache-spark==2.0.0
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-google==6.8.0
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-oracle==2.0.0
apache-airflow-providers-sqlite==2.1.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib